+++ EDF_FileReader/src/edf_filereader/data/ContinuousData.java
@@ -1,7 +1,3 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package edf_filereader.data;
 
 import java.util.List;
@@ -24,10 +20,10 @@ public class ContinuousData {
             List<String> labelsOfTheChannels,
             List<String> transducerTypes,
             List<String> physicalDimensionOfChannels,
-            List<Integer> physicalMinimums,
-            List<Integer> physicalMaximums,
-            List<Integer> digitalMinimums,
-            List<Integer> digitalMaximums,
+            List<Double> physicalMinimums,
+            List<Double> physicalMaximums,
+            List<Double> digitalMinimums,
+            List<Double> digitalMaximums,
             List<String> prefilterings,
             List<Integer> numberOfSamples,
             int sampleLength
